requirements.txt
alembic==0.9.9
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-Scss==0.5
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
PyJWT==1.6.4
pyScss==1.3.5
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.8
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2
